Issue #16: Split domain specific functionality from mytsplugin to adapter.
